+++ src/kernel/GaussianKernel.h
@@ -18,8 +18,8 @@
 class CGaussianKernel: public CRealKernel
 {
  public:
-  CGaussianKernel(LONG size, double width);
-  CGaussianKernel(CRealFeatures* l, CRealFeatures* r, LONG size, double width);
+  CGaussianKernel(INT size, double width);
+  CGaussianKernel(CRealFeatures* l, CRealFeatures* r, INT size, double width);
   ~CGaussianKernel() ;
   
   virtual bool init(CFeatures* l, CFeatures* r, bool do_init);
